requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# Database
sqlalchemy>=2.0.25

# PDF processing
pymupdf>=1.23.21
pymupdf4llm>=0.0.17

# LLM integration
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# Config
python-dotenv>=1.0.0
